Refactor Openstack client
