+++ source/lib/vagrant-openstack-provider/action/resume.rb
@@ -9,8 +9,7 @@ module VagrantPlugins
         def call(env)
           if env[:machine].id
             env[:ui].info I18n.t("vagrant.actions.vm.resume.resuming")
-            client = env[:openstack_client]
-            client.resume_server(env, env[:machine].id)
+            env[:openstack_client].nova.resume_server(env, env[:machine].id)
           end
 
           @app.call(env)
