+++ source/lib/vagrant-openstack-provider/action/wait_stop.rb
@@ -13,7 +13,7 @@ module VagrantPlugins
         def call(env)
           if env[:machine].id
             env[:ui].info(I18n.t("vagrant_openstack.waiting_stop"))
-            client = env[:openstack_client]
+            client = env[:openstack_client].nova
             timeout(200) do
               while client.get_server_details(env, env[:machine].id)['status'] != 'SHUTOFF'
                 sleep 3
