+++ source/lib/vagrant-openstack-provider/client/openstack.rb
@@ -1,0 +1,36 @@
+require "log4r"
+require "restclient"
+require "json"
+
+require "vagrant-openstack-provider/client/keystone"
+require "vagrant-openstack-provider/client/nova"
+
+module VagrantPlugins
+  module Openstack
+    class Session
+      include Singleton
+
+      attr_accessor :token
+      attr_accessor :project_id
+      attr_accessor :endpoints
+
+      def initialize
+        @token = nil
+        @project_id = nil
+        @endpoints = {}
+      end
+    end
+
+    def self.session
+      Session.instance
+    end
+
+    def self.keystone
+      Openstack::KeystoneClient.new
+    end
+
+    def self.nova
+      Openstack::NovaClient.new
+    end
+  end
+end
